@@ -1,9 +1,8 @@
-import React from 'react';
 import LuigiClient from '@luigi-project/client';
 import { useEffect } from 'react';
 
 export function setWindowTitle(title) {
-  setImmediate(() =>
+  setTimeout(() =>
     LuigiClient.sendCustomMessage({ id: 'busola.setWindowTitle', title }),
   );
 }
@@ -12,9 +11,7 @@ export function useWindowTitle(title) {
   useEffect(() => setWindowTitle(title), [title]);
 }
 
-export function withTitle(title, Component) {
-  return props => {
-    setWindowTitle(title);
-    return <Component {...props} />;
-  };
+export function WithTitle({ title, children }) {
+  useWindowTitle(title);
+  return children;
 }
