@@ -1,6 +1,6 @@
 import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import '@testing-library/jest-dom/extend-expect';
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import '@testing-library/jest-dom';
 import { act } from '@testing-library/react';
 
 Enzyme.configure({ adapter: new Adapter() });
