@@ -1,6 +1,6 @@
 const isLocalDev = location.hostname.startsWith('localhost');
 
-export function setupMonaco(monaco) {
+export function setupMonaco(loader) {
   // monaco editor - load from static files instead of from CDN
-  monaco.config({ paths: { vs: isLocalDev ? '/vs' : '/assets/libs/vs' } });
+  loader.config({ paths: { vs: isLocalDev ? '/vs' : '/assets/libs/vs' } });
 }
